renaming mpopt to mdopt because apparently mpopt already existed
