+++ .github/workflows/lint.yml
@@ -20,4 +20,4 @@ jobs:
           run: poetry install
 
         - name: Run linter
-          run: poetry run pylint mpopt --fail-under=9
+          run: poetry run pylint mdopt --fail-under=9
