+++ .github/workflows/tests.yml
@@ -28,5 +28,5 @@ jobs:
         with:
           fail_ci_if_error: true
           token: ${{ secrets.CODECOV_TOKEN }}
-          files: /home/runner/work/mpopt/mpopt/coverage.xml
+          files: /home/runner/work/mdopt/mdopt/coverage.xml
           verbose: true
