--- CNAME
@@ -1,1 +1,0 @@
-mpopt.readthedocs.io
