+++ README.rst
@@ -1,6 +1,6 @@
-# mpopt
+# mdopt
 MPS + MPO optimiser library
 
-[![codecov](https://codecov.io/gh/quicophy/mpopt/branch/main/graph/badge.svg?token=4G7VWYX0S2)](https://codecov.io/gh/quicophy/mpopt)
-[![tests](https://github.com/quicophy/mpopt/actions/workflows/tests.yml/badge.svg?branch=main)](https://github.com/quicophy/mpopt/actions/workflows/tests.yml)
+[![codecov](https://codecov.io/gh/quicophy/mdopt/branch/main/graph/badge.svg?token=4G7VWYX0S2)](https://codecov.io/gh/quicophy/mdopt)
+[![tests](https://github.com/quicophy/mdopt/actions/workflows/tests.yml/badge.svg?branch=main)](https://github.com/quicophy/mdopt/actions/workflows/tests.yml)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
