+++ experiments/ising_parallel.py
@@ -21,24 +21,24 @@ from scipy.sparse import csr_matrix, eye, kron
 from scipy.sparse.linalg import eigsh
 
 try:
-    from mpopt.mps.explicit import ExplicitMPS
-    from mpopt.mps.canonical import CanonicalMPS
-    from mpopt.optimiser.dmrg import DMRG as dmrg_optimiser
-    from mpopt.contractor.contractor import (
+    from mdopt.mps.explicit import ExplicitMPS
+    from mdopt.mps.canonical import CanonicalMPS
+    from mdopt.optimiser.dmrg import DMRG as dmrg_optimiser
+    from mdopt.contractor.contractor import (
         apply_one_site_operator,
         apply_two_site_unitary,
     )
-    from mpopt.mps.utils import inner_product, create_simple_product_state
+    from mdopt.mps.utils import inner_product, create_simple_product_state
 except ImportError:
     sys.path[0] += "/.."
-    from mpopt.mps.explicit import ExplicitMPS
-    from mpopt.mps.canonical import CanonicalMPS
-    from mpopt.optimiser.dmrg import DMRG as dmrg_optimiser
-    from mpopt.contractor.contractor import (
+    from mdopt.mps.explicit import ExplicitMPS
+    from mdopt.mps.canonical import CanonicalMPS
+    from mdopt.optimiser.dmrg import DMRG as dmrg_optimiser
+    from mdopt.contractor.contractor import (
         apply_one_site_operator,
         apply_two_site_unitary,
     )
-    from mpopt.mps.utils import inner_product, create_simple_product_state
+    from mdopt.mps.utils import inner_product, create_simple_product_state
 
 
 def compute_one_site_expectation_value(
@@ -235,7 +235,7 @@ class IsingMPO:
     def hamiltonian_mpo(self) -> list[np.ndarray]:
         """Returns a Matrix Product Operator representation of the Hamiltonian.
 
-        Follows the convention of indices from ::module:: `mpopt.mps.explicit.py`:
+        Follows the convention of indices from ::module:: `mdopt.mps.explicit.py`:
         each tensor in the MPO list has legs (vL, vR, pU, pD),
         where v stands for "virtual", p -- for "physical",
         and L, R, U, D stand for "left", "right", "up", "down".
