+++ mdopt/mps/__init__.py
@@ -1,0 +1,1 @@
+from .explicit import ExplicitMPS
