+++ mdopt/optimiser/__init__.py
@@ -1,0 +1,1 @@
+from .dmrg import DMRG
