+++ mdopt/optimiser/dmrg.py
@@ -1,28 +1,6 @@
-"""This module contains the :class:`DephasingDMRG` and the :class:`EffectiveDensityOperator`
-classes.
-
-This algorithm's main feature is that it restricts the target-state search to
-the computational basis states domain.
-In particular, we use it to find the main component of a Matrix Density Product Operator (MDPO),
-i.e., a computational basis state contributing the largest amplitude.
-
-In our notation, MDPO for `n` sites denotes the following object.
-
-```
-     |      |               |       |
-     |      |               |       |
-----(0*)---(1*)--- ... ---(n-2*)--(n-1*)---
-
-----(0)----(1)---- ... ---(n-2)---(n-1)----
-     |      |               |       |
-     |      |               |       |
-```
-
-An MDPO formed by an MPS and its complex-conjugated version.
-The main idea is to find the main component of this object without
-performing the kronecker product explicitly.
+"""This module contains the :class:`DMRG` and the :class:`EffectiveOperator` classes.
+Inspired by TenPy.
 """
-
 
 from typing import Union
 import numpy as np
@@ -31,89 +9,103 @@ from opt_einsum import contract
 from scipy.sparse.linalg import eigsh
 from tqdm import tqdm
 
-from mpopt.mps.canonical import CanonicalMPS
-from mpopt.mps.explicit import ExplicitMPS
-from mpopt.utils.utils import split_two_site_tensor
+from mdopt.mps.canonical import CanonicalMPS
+from mdopt.mps.explicit import ExplicitMPS
+from mdopt.utils.utils import split_two_site_tensor
 
 
-class EffectiveDensityOperator(scipy.sparse.linalg.LinearOperator):
-    r"""Class to store an effective two-site density operator.
+class EffectiveOperator(scipy.sparse.linalg.LinearOperator):
+    r"""Class to store an effective two-site operator.
 
-    To take more advantage of :module:`scipy.sparse.linalg`, we make a special class
-    for local effective density operators extending the analogy from local effective operators.
+    In order to take more advantage of the :module:`scipy.sparse.linalg` module,
+    we make a special class for local effective operators.
     It allows us to compute eigenvectors more effeciently.
 
-    The diagram displaying the contraction can be found in the supplementary notes.
+    Such effective operator is to be diagonalised in the
+    :method:`update_bond` method of the :class:'DMRG' class.
+
+    ```
+     ---uL                    uR---
+     |        i          j        |
+     |  vL    |    d     |    vR  |
+    (L)----(mpo_l)----(mpo_r)----(R)
+     |        |          |        |
+     |        k          l        |
+     ---dL                    dR---
+     ```
+
+    In our convention, the legs of left/right environments (tensors `L`/`R` in the cartoon)
+    are ordered as follows: (`uL`/`uR`, `vL`/`vR`, `dL`/`dR`) which means (up, virtual, down).
     """
 
     def __init__(
         self,
         left_environment: np.ndarray,
-        mps_target_1: np.ndarray,
-        mps_target_2: np.ndarray,
+        mpo_tensor_left: np.ndarray,
+        mpo_tensor_right: np.ndarray,
         right_environment: np.ndarray,
     ):
-        """
-        Initialise an effective dephased density operator tensor network.
+        """Initialises an effective operator tensor network.
 
         Parameters
             left_environment :
-                The left environment for the effective dephased density operator.
-            mps_target_1 :
-                The left target matrix product state tensor.
-            mps_target_2 :
-                The right target matrix product state tensor.
-            right_environment :
-                The right environment for the effective dephased density operator.
+                The left environment for the effective operator.
+            mpo_tensor_left :
+                The left MPO tensor.
+            mpo_tensor_right :
+                The right MPO tensor.
+            right_environment:
+                The right environment for the effective operator.
         """
-        if len(left_environment.shape) != 4:
+
+        if len(left_environment.shape) != 3:
             raise ValueError(
-                "A valid left environment tensor must have 4 legs"
+                "A valid left environment tensor must have 3 legs"
                 f"while the one given has {len(left_environment.shape)}."
             )
-        if len(mps_target_1.shape) != 3:
+        if len(mpo_tensor_left.shape) != 4:
             raise ValueError(
-                "A valid target MPS tensor must have 3 legs"
-                f"while the one given has {len(mps_target_1.shape)}."
+                "A valid mpo left tensor must have 4 legs"
+                f"while the one given has {len(mpo_tensor_left.shape)}."
             )
-        if len(mps_target_2.shape) != 3:
+        if len(mpo_tensor_right.shape) != 4:
             raise ValueError(
-                "A valid target MPS tensor must have 3 legs"
-                f"while the one given has {len(mps_target_1.shape)}."
+                "A valid mpo right tensor must have 4 legs"
+                f"while the one given has {len(mpo_tensor_right.shape)}."
             )
-        if len(right_environment.shape) != 4:
+        if len(right_environment.shape) != 3:
             raise ValueError(
-                "A valid right environment tensor must have 4 legs"
+                "A valid right environment tensor must have 3 legs"
                 f"while the one given has {len(right_environment.shape)}."
             )
 
         self.left_environment = left_environment
         self.right_environment = right_environment
-        self.mps_target_1 = mps_target_1
-        self.mps_target_2 = mps_target_2
+        self.mpo_tensor_left = mpo_tensor_left
+        self.mpo_tensor_right = mpo_tensor_right
         chi_1, chi_2 = (
-            left_environment.shape[3],
-            right_environment.shape[3],
+            left_environment.shape[2],
+            right_environment.shape[2],
         )
         d_1, d_2 = (
-            mps_target_1.shape[1],
-            mps_target_2.shape[1],
+            mpo_tensor_left.shape[3],
+            mpo_tensor_right.shape[3],
         )
         self.x_shape = (chi_1, d_1, d_2, chi_2)
         self.shape = (chi_1 * d_1 * d_2 * chi_2, chi_1 * d_1 * d_2 * chi_2)
-        self.dtype = mps_target_1.dtype
+        self.dtype = mpo_tensor_left.dtype
         super().__init__(shape=self.shape, dtype=self.dtype)
 
     def _matvec(self, x: np.ndarray) -> np.ndarray:
         """Performs matrix-vector multiplication.
 
-        Computes effective_density_operator * |x> = |x'>
-        This function is used by :func:`scipy.sparse.linalg.eigsh` to diagonalise
-        the effective density operator with the Lanczos method, withouth generating the full matrix.
+        Computes effective_operator * |x> = |x'>.
+        This function is being used by :func:`scipy.sparse.linalg.eigsh` to diagonalise
+        the effective operator with the Lanczos method, without generating the full matrix.
 
         Parameters
-            x : np.array
-                The two-site tensor we are acting on with an effective density operator.
+            x :
+                The two-site tensor on which acts an effective operator.
         """
 
         two_site_tensor = np.reshape(x, self.x_shape)
@@ -124,55 +116,34 @@ class EffectiveDensityOperator(scipy.sparse.linalg.LinearOperator):
                 f"while the one given has {len(two_site_tensor.shape)}."
             )
 
-        copy_tensor = np.fromfunction(
-            lambda i, j, k: np.logical_and(i == j, j == k), (2, 2, 2), dtype=np.int16
-        )
-
-        einsum_string = (
-            "ustw, ailu, ifj, jhk, bef, cgh, lom, mpn, eso, gtp, dknw -> abcd"
-        )
+        einsum_string = "ijkl, mni, nopj, oqrk, sql -> mprs"
         two_site_tensor = contract(
             einsum_string,
             two_site_tensor,
             self.left_environment,
-            np.conjugate(self.mps_target_1),
-            np.conjugate(self.mps_target_2),
-            copy_tensor,
-            copy_tensor,
-            self.mps_target_1,
-            self.mps_target_2,
-            copy_tensor,
-            copy_tensor,
+            self.mpo_tensor_left,
+            self.mpo_tensor_right,
             self.right_environment,
-            optimize=[
-                (0, 8),
-                (0, 1),
-                (0, 6),
-                (0, 5),
-                (1, 2),
-                (2, 3),
-                (3, 4),
-                (2, 3),
-                (1, 2),
-                (0, 1),
-            ],
+            optimize=[(0, 1), (0, 3), (0, 2), (0, 1)],
         )
 
         return np.reshape(two_site_tensor, self.shape[0])
 
 
-class DephasingDMRG:
-    """Class storing the Dephasing DMRG methods.
+class DMRG:
+    """Class storing the DMRG methods.
 
-    Class holding the Dephasing Density Matrix Renormalisation Group algorithm with two-site updates
+    Class holding the Density Matrix Renormalisation Group algorithm with two-site updates
     for a finite-size system with open-boundary conditions.
 
     Attributes:
         mps :
             MPS serving as a current approximation of the target state.
-        mps_target :
-            The "target" MPS in the right-canonical form.
-            This MPS is used to construct the dephased MDPO.
+        mpo :
+            The MPO of which the target state is to be computed.
+            Each tensor in the MPO list has legs `(vL, vR, pU, pD)`,
+            where `v` stands for "virtual", `p` -- for "physical",
+            and `L`, `R`, `U`, `D` -- for "left", "right", "up", "down" accordingly.
         chi_max :
             The highest bond dimension of an MPS allowed.
         mode :
@@ -191,54 +162,56 @@ class DephasingDMRG:
     def __init__(
         self,
         mps: Union[ExplicitMPS, CanonicalMPS],
-        mps_target: Union[ExplicitMPS, CanonicalMPS],
+        mpo: list[np.ndarray],
         chi_max: np.int16 = 1e4,
         cut: np.float64 = 1e-12,
         mode: str = "SA",
         silent: bool = False,
         copy: bool = True,
     ):
-        if len(mps) != len(mps_target):
+
+        if len(mps) != len(mpo):
             raise ValueError(
                 f"The MPS has length {len(mps)},"
-                f"the target MPS has length {len(mps_target)},"
+                f"the MPO has length {len(mpo)},"
                 "but the lengths should be equal."
             )
+        for i, tensor in enumerate(mpo):
+            if len(tensor.shape) != 4:
+                raise ValueError(
+                    f"A valid MPO tensor must have 4 legs while tensor {i} has {len(tensor.shape)}."
+                )
+        if mode not in ["SA", "LA", "SM", "LM"]:
+            raise ValueError("Invalid eigensolver mode given.")
+
+        self.mps = mps
         if copy:
             self.mps = mps.copy()
         if isinstance(self.mps, CanonicalMPS):
             self.mps = self.mps.right_canonical()
-        self.mps = mps
         self.left_environments = [None] * len(mps)
         self.right_environments = [None] * len(mps)
-        mps_target = mps_target.right_canonical()
-        self.mps_target = mps_target
+        self.mpo = mpo
         self.chi_max = chi_max
         self.cut = cut
         self.mode = mode
         self.silent = silent
 
-        start_bond_dim = self.mps_target.tensors[0].shape[0]
+        start_bond_dim = self.mpo[0].shape[0]
         chi = mps.tensors[0].shape[0]
-        left_environment = np.zeros(
-            [chi, start_bond_dim, start_bond_dim, chi], dtype=np.float64
-        )
-        right_environment = np.zeros(
-            [chi, start_bond_dim, start_bond_dim, chi], dtype=np.float64
-        )
-        left_environment[:, 0, 0, :] = np.eye(chi, dtype=np.float64)
-        right_environment[:, start_bond_dim - 1, start_bond_dim - 1, :] = np.eye(
-            chi, dtype=np.float64
-        )
+        left_environment = np.zeros([chi, start_bond_dim, chi], dtype=np.float64)
+        right_environment = np.zeros([chi, start_bond_dim, chi], dtype=np.float64)
+        left_environment[:, 0, :] = np.eye(chi, dtype=np.float64)
+        right_environment[:, start_bond_dim - 1, :] = np.eye(chi, dtype=np.float64)
         self.left_environments[0] = left_environment
         self.right_environments[-1] = right_environment
         for i in reversed(range(1, len(mps))):
             self.update_right_environment(i)
 
     def sweep(self):
-        """One Dephasing DMRG sweep.
+        """One DMRG sweep.
 
-        A method performing one Dephasing DMRG sweep, which consists of
+        A method performing one DMRG sweep, which consists of
         two series of `update_bond` sweeps which go back and forth.
         """
 
@@ -253,39 +226,39 @@ class DephasingDMRG:
 
         j = i + 1
 
-        effective_density_operator = EffectiveDensityOperator(
+        effective_hamiltonian = EffectiveOperator(
             self.left_environments[i],
-            self.mps_target.tensors[i],
-            self.mps_target.tensors[j],
+            self.mpo[i],
+            self.mpo[j],
             self.right_environments[j],
         )
 
         if isinstance(self.mps, ExplicitMPS):
             initial_guess = self.mps.two_site_right_iso(i).reshape(
-                effective_density_operator.shape[0]
+                effective_hamiltonian.shape[0]
             )
         if isinstance(self.mps, CanonicalMPS):
             self.mps = self.mps.move_orth_centre(i)
             initial_guess = self.mps.two_site_tensor_next(i).reshape(
-                effective_density_operator.shape[0]
+                effective_hamiltonian.shape[0]
             )
 
         _, eigenvectors = eigsh(
-            effective_density_operator,
+            effective_hamiltonian,
             k=1,
             which=self.mode,
             return_eigenvectors=True,
             v0=initial_guess,
             tol=1e-8,
         )
-        x = eigenvectors[:, 0].reshape(effective_density_operator.x_shape)
+        x = eigenvectors[:, 0].reshape(effective_hamiltonian.x_shape)
         left_iso_i, singular_values_j, right_iso_j = split_two_site_tensor(
             x, chi_max=self.chi_max, cut=self.cut, renormalise=True
         )
 
         if isinstance(self.mps, CanonicalMPS):
             self.mps.tensors[i] = np.tensordot(
-                left_iso_i, np.diag(singular_values_j), (1, 0)
+                left_iso_i, np.diag(singular_values_j), (2, 0)
             )
             self.mps.orth_centre = i
             self.mps.tensors[j] = right_iso_j
@@ -318,13 +291,12 @@ class DephasingDMRG:
             right_iso = self.mps.single_site_tensor(i)
 
         right_environment = contract(
-            "ijkl, omi, pmj, qnk, rnl -> opqr",
-            right_environment,
+            "ijk, lnjm, omp, knp -> ilo",
             right_iso,
-            np.conjugate(self.mps_target.tensors[i]),
-            self.mps_target.tensors[i],
+            self.mpo[i],
             np.conjugate(right_iso),
-            optimize=[(0, 2), (0, 1), (0, 1), (0, 1)],
+            right_environment,
+            optimize=[(0, 3), (0, 2), (0, 1)],
         )
         self.right_environments[i - 1] = right_environment
 
@@ -342,13 +314,12 @@ class DephasingDMRG:
             left_iso = self.mps.single_site_tensor(i)
 
         left_environment = contract(
-            "ijkl, imo, jmp, knq, lnr -> opqr",
-            left_environment,
+            "ijk, lnjm, omp, ilo -> knp",
             left_iso,
-            np.conjugate(self.mps_target.tensors[i]),
-            self.mps_target.tensors[i],
+            self.mpo[i],
             np.conjugate(left_iso),
-            optimize=[(0, 2), (0, 1), (0, 1), (0, 1)],
+            left_environment,
+            optimize=[(0, 3), (0, 2), (0, 1)],
         )
         self.left_environments[i + 1] = left_environment
