--- mpopt/mps/__init__.py
@@ -1,3 +1,0 @@
-"""This module contains the Explicit and Canonical MPS constructions along with the relevant functions."""
-
-from .explicit import ExplicitMPS
