--- mpopt/optimiser/__init__.py
@@ -1,3 +1,0 @@
-"""This module contains optimisers for matrix product states."""
-
-from .dmrg import DMRG
